remove util deps from testcommon so we can use in util tests
