+++ pkg/testcommon/testcommon.go
@@ -2,7 +2,6 @@ package testcommon
 
 import (
 	"bytes"
-	"errors"
 	"io"
 	"io/ioutil"
 	"math/rand"
@@ -11,7 +10,6 @@ import (
 	"time"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/drud/ddev/pkg/util"
 
 	"github.com/drud/ddev/pkg/ddevapp"
 	"github.com/drud/drud-go/utils/system"
@@ -33,9 +31,6 @@ type TestSite struct {
 	Dir string
 }
 
-// NetName provides the default network name for testing.
-const NetName = "ddev_default"
-
 func (site *TestSite) createArchivePath() string {
 	dir := CreateTmpDir(site.Name + "download")
 	return filepath.Join(dir, site.Name+".tar.gz")
@@ -47,7 +42,7 @@ func (site *TestSite) Prepare() error {
 	site.Dir = testDir
 	log.Debugf("Prepping test for %s.\n", site.Name)
 	err := os.Setenv("DRUD_NONINTERACTIVE", "true")
-	util.CheckErr(err)
+	checkErr(err)
 
 	log.Debugln("Downloading file:", site.SourceURL)
 	site.ArchivePath = site.createArchivePath()
@@ -102,7 +97,7 @@ func (site *TestSite) Chdir() func() {
 // Cleanup removes the archive and codebase extraction for a site after a test run has completed.
 func (site *TestSite) Cleanup() {
 	err := os.Remove(site.ArchivePath)
-	util.CheckErr(err)
+	checkErr(err)
 	// CleanupDir checks its own errors.
 	CleanupDir(site.Dir)
 }
@@ -110,7 +105,7 @@ func (site *TestSite) Cleanup() {
 // CleanupDir removes a directory specified by string.
 func CleanupDir(dir string) {
 	err := os.RemoveAll(dir)
-	util.CheckErr(err)
+	checkErr(err)
 }
 
 // OsTempDir gets os.TempDir() (usually provided by $TMPDIR) but expands any symlinks found within it.
@@ -190,12 +185,12 @@ func CaptureStdOut() func() string {
 		go func() {
 			var buf bytes.Buffer
 			_, err := io.Copy(&buf, r)
-			util.CheckErr(err)
+			checkErr(err)
 			outC <- buf.String()
 		}()
 
 		// back to normal state
-		util.CheckClose(w)
+		checkClose(w)
 		os.Stdout = old // restoring the real stdout
 		out := <-outC
 		return out
@@ -224,29 +219,18 @@ func ClearDockerEnv() {
 	}
 }
 
-// ContainerCheck determines if a given container name exists and matches a given state
-func ContainerCheck(checkName string, checkState string) (bool, error) {
-	// ensure we have docker network
-	client := util.GetDockerClient()
-	err := util.EnsureNetwork(client, NetName)
+// checkErr exits with a log.Fatal() if an error is encountered.
+// It is normally used for errors that we never expect to happen, and don't have any normal handling technique.
+// From https://davidnix.io/post/error-handling-in-go/
+func checkErr(err error) {
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("ERROR:", err)
 	}
+}
 
-	containers, err := util.GetDockerContainers(true)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	for _, container := range containers {
-		name := util.ContainerName(container)
-		if name == checkName {
-			if container.State == checkState {
-				return true, nil
-			}
-			return false, errors.New("container " + name + " returned " + container.State)
-		}
-	}
-
-	return false, errors.New("unable to find container " + checkName)
+// checkClose is used to check the return from Close in a defer statement.
+// From https://groups.google.com/d/msg/golang-nuts/-eo7navkp10/BY3ym_vMhRcJ
+func checkClose(c io.Closer) {
+	err := c.Close()
+	checkErr(err)
 }
